chore(pectra): Update geth dependencies for Pectra. (#13)
## 📝 Summary

We need to use at least geth 1.15.0 for Pectra, since that's the version
that can parse tx type 4 (the new EIP-7702 tx type).

The code doing so is in go-utils so we can either wait for a new tagged
release that includes https://github.com/flashbots/go-utils/pull/37 or
update directly like in this PR.

## ⛱ Motivation and Context

Pectra.

## 📚 References

<!-- Any interesting external links to documentation, articles, tweets
which add value to the PR -->

---

## ✅ I have run these commands

* [x] `make lint`
* [x] `make test`
* [x] `go mod tidy`