Fix release CI (#56)
## 📝 Summary

Current Release CI is broken. I have added a simple one to build and
upload binaries.
Right now it doesn't run automatically on every tag, but needs to be
triggered manually.

I have tested in a fork. 

This is a test run:
https://github.com/niccoloraspa/buildernet-orderflow-proxy/actions/runs/18122690530
 
This is the resulting release:
https://github.com/niccoloraspa/buildernet-orderflow-proxy/releases/tag/v0.0.1


I have removed the Docker building/pushing part from the release CI as
it wasn't working. I can fix that in another PR if needed.

## ✅ I have run these commands

* [x] `make lint`
* [x] `make test`
* [x] `go mod tidy`